requirements.txt
channels
django-filter
djangorestframework
pillow